jpa repo/service implemented. dao/service/controller commented out
push before deleting previous code

+++ src/main/java/com/springboot/titletech/StageTestData.java
@@ -108,7 +108,7 @@ public class StageTestData {
 
         for (ParcelOwnership p : parcelOwnershipList) {
             if(p.getParcelid() == parcelDocumentList.get(idToCheck).getParcelid()){
-                parcelOwnershipid = parcelDocumentList.get(idToCheck).getId();
+                parcelOwnershipid = (int) parcelDocumentList.get(idToCheck).getId();
                 break;
             }
         }
